requirements.txt
streamlit>=1.32
anthropic>=0.25
httpx>=0.27
python-docx>=1.1
mammoth>=1.7
python-dotenv>=1.0
